Adjusted the bounce stun time

+++ Assets/Gameplay/Enemies/ShieldEnemy.cs
@@ -8,7 +8,7 @@ public class ShieldEnemy : MonoBehaviour
 
     public Transform PlayerPosition;
     public int AggressionRange;
-    public int StoppingDistance;
+    public float StoppingDistance;
     
     //###############
     //##  MEMBERS  ##
